feat: altera tipo de ID de Long para Integer nas notificações
- Modifica o tipo de ID nas classes NotificacaoController, NotificacaoService e NotificacaoRepository de Long para Integer, garantindo consistência no tratamento de identificadores.
- Essas alterações visam simplificar a manipulação de IDs e melhorar a integração com o banco de dados, mantendo a integridade das operações de CRUD nas notificações.

+++ src/main/java/controller/NotificacaoController.java
@@ -20,7 +20,7 @@ public class NotificacaoController {
     }
 
     @GetMapping("/entidade/{entidadeId}")
-    public String listarNotificacoesPorEntidade(@PathVariable Long entidadeId, Model model) {
+    public String listarNotificacoesPorEntidade(@PathVariable Integer entidadeId, Model model) {
         model.addAttribute("notificacoes", notificacaoService.listarPorEntidade(entidadeId));
         return "notificacao/lista";
     }
@@ -32,27 +32,27 @@ public class NotificacaoController {
     }
 
     @GetMapping("/entidade/{entidadeId}/nao-lidas")
-    public String listarNotificacoesNaoLidasPorEntidade(@PathVariable Long entidadeId, Model model) {
+    public String listarNotificacoesNaoLidasPorEntidade(@PathVariable Integer entidadeId, Model model) {
         model.addAttribute("notificacoes", notificacaoService.listarNaoLidasPorEntidade(entidadeId));
         return "notificacao/lista";
     }
 
     @PostMapping("/{id}/marcar-lida")
     @ResponseBody
-    public String marcarComoLida(@PathVariable Long id) {
+    public String marcarComoLida(@PathVariable Integer id) {
         notificacaoService.marcarComoLida(id);
         return "Notificação marcada como lida";
     }
 
     @PostMapping("/entidade/{entidadeId}/marcar-todas-lidas")
     @ResponseBody
-    public String marcarTodasComoLidas(@PathVariable Long entidadeId) {
+    public String marcarTodasComoLidas(@PathVariable Integer entidadeId) {
         notificacaoService.marcarTodasComoLidas(entidadeId);
         return "Todas as notificações foram marcadas como lidas";
     }
 
     @GetMapping("/{id}/deletar")
-    public String deletarNotificacao(@PathVariable Long id) {
+    public String deletarNotificacao(@PathVariable Integer id) {
         notificacaoService.deletar(id);
         return "redirect:/notificacoes";
     }
